requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
rich>=13.0.0
asyncio-throttle>=1.0.0

# File system and archive handling
py7zr>=0.20.0
rarfile>=4.0
python-magic>=0.4.27
patool>=1.12.0

# Security analysis tools
yara-python>=4.3.0
capstone>=5.0.0
keystone-engine>=0.9.2
unicorn>=2.0.0

# Web and reporting
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
aiofiles>=23.0.0

# Database and data handling
sqlalchemy>=2.0.0
alembic>=1.12.0
pandas>=2.0.0

# System and process management
psutil>=5.9.0
docker>=6.1.0
subprocess32>=3.5.4

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

